Adding a conditional statement
Adding a conditional statement, in case it has been installed before

+++ scripts/install-empire.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
-cd /opt
-git clone https://github.com/EmpireProject/Empire.git
-cd Empire
-expect -c 'spawn ./setup/install.sh ; expect "server"; send "\r"; interact'
-ln -s empire /usr/bin/empire
+FILE=/opt/Empire/empire
+if [ -f "$FILE" ]; then
+    echo "Empire has been installed"
+else
+	cd /opt
+	git clone https://github.com/EmpireProject/Empire.git
+	cd Empire
+	expect -c 'spawn ./setup/install.sh ; expect "server"; send "\r"; interact'
+	ln -s empire /usr/bin/empire
+fi
